\name{manifoldsData-package}
\alias{manifoldsData}
\docType{package}
\title{manifolds data collection}
\description{Collecting manifolds data in the textbook "Nonparametric Statistics on Manifolds and Their Applications to Object Data Analysis"}
\details{
\tabular{ll}{
Package: \tab manifoldsData\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2018-9-10\cr
License: \tab GPL-3\cr
}
}

\author{
Jin Zhu

Maintainer: Jin Zhu <zhuj37@mail2.sysu.edu.cn>
}

