% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orbit}
\alias{orbit}
\title{orbit planet data}
\format{\itemize{
\code{nx}: the first coordinates of the unit normals to orbital planes

\code{ny}: the second coordinates of the unit normals to orbital planes

\code{ny}: the third coordinates of the unit normals to orbital planes
}}
\description{
A orbit planet data include 9 records about the unit normals to orbital planes.
}
\details{
This orbit planet data containing 9 records about the unit normals to orbital planes.
The unit normals is (sin(a)sin(b), -cos(a)sin(b), cos(b)) where a is the the inclination of
the orbital plane of a planet to the ecliptic and b is the angle between a fixed line in the ecliptic
and the line joining the ascending node of the planet.
}
