% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skullnormal}
\alias{skullnormal}
\title{skull data}
\format{\itemize{
\code{skullnormal$x}: A data frame include the landmark of skull

\code{skullnormal$y}: A data frame include the demographics such as months and gender.
}}
\description{
A skull data include 124 records about the landmark of skull and demographics of healthy samples.
}
\references{
Patrangenaru, Victor, and Leif Ellingson. Nonparametric statistics on manifolds and their applications to object data analysis. CRC Press, 2015.
}
